Add support for Java's native Normalizer.
The Unicode gem compiles on JRuby as a C extension, which is bad
for performance (and won't run, for example, on Travis-CI).
Detect when we're running under JRuby and switch to using Java's
native Normalizer.  Also, Java's normalizer does not normalize a
floating orphan diacritic tilde into a proper tilde (nor, I think,
should it).  Support \textasciitilde instead.